Modificación en Elasticsearc

+++ JAVA/elastic-example/src/main/java/com/chuidiang/Product.java
@@ -1,4 +1,4 @@
 package com.chuidiang;
 
-public record Product(String id, String name, double value) {
+public record Product(String id, String name, Integer value) {
 }
